+++ Assets/XFramework/Tests/Runtime/Node/UpdateNodeIntegrationTests.cs
@@ -208,7 +208,7 @@ namespace XFramework.XUpdate.Tests
             public UpdateLOD OnUpdate(float deltaTime, float time)
             {
                 OnUpdateCallCount++;
-                return UpdateLOD.Frame1;
+                return UpdateLOD.Tier0;
             }
         }
 
@@ -250,7 +250,7 @@ namespace XFramework.XUpdate.Tests
             public UpdateLOD OnFixedUpdate(float deltaTime, float fixedTime)
             {
                 OnFixedUpdateCallCount++;
-                return UpdateLOD.Frame1;
+                return UpdateLOD.Tier0;
             }
         }
 
@@ -276,7 +276,7 @@ namespace XFramework.XUpdate.Tests
             public UpdateLOD OnLateUpdate(float deltaTime, float time)
             {
                 OnLateUpdateCallCount++;
-                return UpdateLOD.Frame1;
+                return UpdateLOD.Tier0;
             }
         }
 
@@ -303,7 +303,7 @@ namespace XFramework.XUpdate.Tests
             public UpdateLOD OnUpdate(float deltaTime, float time)
             {
                 OnUpdateCallCount++;
-                return UpdateLOD.Frame1;
+                return UpdateLOD.Tier0;
             }
         }
     }